Attempt at Inversify, thought it's not working
