+++ jest.config.js
@@ -4,5 +4,9 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  moduleDirectories: ['node_modules', 'src']
+  moduleDirectories: ['node_modules', 'src'],
+  transform: {
+    "\\.sh$": "<rootDir>/src/test/jest-raw-loader.js",
+     "\\.yml$": "<rootDir>/src/test/jest-raw-loader.js"
+  }
 };
