+++ src/common/FakeFsGateway.ts
@@ -1,0 +1,13 @@
+import { decorate, injectable } from 'inversify';
+
+/** Stub class for filesystem. */
+class FakeFsGateway {
+  fs = {
+    existsSync: () => {},
+    mkdirSync: () => {},
+    writeFileSync: () => {}
+  };
+}
+
+decorate(injectable(), FakeFsGateway);
+export default FakeFsGateway;
