+++ src/common/types.ts
@@ -1,0 +1,3 @@
+export default {
+  IFsGateway: Symbol.for('IFsGateway')
+}
