+++ src/test/helpers/index.ts
@@ -2,11 +2,6 @@ import * as path from 'path';
 
 export function getTestFilePath(basename: string, file: string): string {
   return path.resolve(
-    __dirname,
-    '..',
-    '..',
-    '..',
-    'src',
     'test',
     basename,
     file
