+++ src/test/suite/utils/getAllJobs.test.ts
@@ -2,7 +2,7 @@ import * as assert from 'assert';
 import * as fs from 'fs';
 import getConfig from 'config/getConfig';
 import getAllJobs from 'job/getJobs';
-import { getTestFilePath } from 'test/helpers/';
+import { getTestFilePath } from 'test/helpers';
 
 suite('getAllJobs', () => {
   test('jobs from fixture', () => {
