+++ src/test/suite/utils/getAttachWorkspaceCommand.test.ts
@@ -1,5 +1,5 @@
 import * as assert from 'assert';
-import { normalize } from 'test/helpers/';
+import { normalize } from 'test/helpers';
 import getAttachWorkspaceCommand from 'config/getAttachWorkspaceCommand';
 
 suite('getAttachWorkspaceCommand', () => {
