+++ tsconfig.json
@@ -15,7 +15,6 @@
   },
   "exclude": [
     "node_modules",
-    ".vscode-test",
     "node"
   ]
 }
